+++ script/actions/SwapSlots.s.sol
@@ -5,8 +5,8 @@ import "./base/SwapSlotsBase.s.sol";
 
 contract SwapSlotsScript is SwapSlotsBaseScript {
     address constant DELEGATOR = 0x0000000000000000000000000000000000000000;
-    uint64 constant INDEX1 = 0;
-    uint64 constant INDEX2 = 0;
+    uint32 constant INDEX1 = 0;
+    uint32 constant INDEX2 = 0;
 
     function run() public {
         runBase(DELEGATOR, INDEX1, INDEX2);
